feat(User sign up): Set up user sign up database >- Set up database user  sign up >- Performs validation with joi >-Writes unit test [delivers #166791008]
