+++ .env
@@ -1,2 +1,3 @@
 secret = thejust
 DB_CONNECTIONSTRING=postgresql://postgres:thejust@localhost:5432/Auto-Mart
+DB_TEST_URL = postgresql://postgres:thejust@localhost:5432/Auto-Mart-Test
