+++ .travis.yml
@@ -12,7 +12,7 @@ services:
   
 before_script:
   - psql -c 'create database test_db;' -U postgres
-  -npm run create
+  - npm run create
 
 script:
   - npm run test
