+++ server/config/config.js
@@ -2,7 +2,15 @@ import { Pool } from 'pg';
 import dotenv from 'dotenv';
 
 dotenv.config();
+
+const connectionstring = {
+  test: process.env.DB_TEST_URL,
+  production: process.env.DB_CONNECTIONSTRING,
+};
+
+const env = process.env.NODE_ENV || 'production';
+
 const pool = new Pool({
-  connectionString: process.env.DB_CONNECTIONSTRING,
+  connectionString: connectionstring[env],
 });
 export default pool;
