+++ server/helpers/schema.js
@@ -7,41 +7,7 @@ const schema = {
     email: joi.string().email().required(),
     password: joi.string().alphanum().min(6).max(8)
       .required(),
-  }),
-
-  signin: joi.object().keys({
-    email: joi.string().email().required(),
-    password: joi.string().alphanum().min(6).max(8)
-      .required(),
-  }),
-
-  postAd: joi.object().keys({
-    manufacturer: joi.string().required(),
-    state: joi.string().required(),
-    model: joi.string().required(),
-    price: joi.number().required(),
-  }),
-
-  purchaseOrder: joi.object().keys({
-    priceOffered: joi.number().required(),
-  }),
-
-  updateOrder: joi.object().keys({
-    newPriceOffered: joi.number().required(),
-  }),
-
-  updateAd: joi.object().keys({
-    newPrice: joi.number().required(),
-  }),
-
-  viewACar: joi.object().keys({
-    carId: joi.number().required(),
-  }),
-
-  viewCars: joi.object().keys({
-    status: joi.string().valid('available'),
-    minPrice: joi.number(),
-    maxPrice: joi.number(),
+    address: joi.string().required(),
   }),
 };
