--- server/model/car.js
@@ -1,70 +1,0 @@
-const car = [
-  {
-    id: 1,
-    owner: 1, 
-    createdOn: new Date(),
-    state: 'new',
-    status: 'available',
-    price: 2000000,
-    manufacturer: 'Toyota',
-    model: 'pathfinder',
-    bodyType: 'car'
-  },
-  {
-    id: 2,
-    owner: 2, 
-    createdOn: new Date(),
-    state: 'new',
-    status: 'available',
-    price: 2500000,
-    manufacturer: 'Toyota',
-    model: 'hilux',
-    bodyType: 'truck'
-  },
-  {
-    id: 3,
-    owner: 3,
-    createdOn: new Date(),
-    state: 'new',
-    status: 'available',
-    price: 2500000,
-    manufacturer: 'Toyota',
-    model: 'hilux',
-    bodyType: 'truck'
-  },
-  {
-    id: 4,
-    owner: 4, 
-    createdOn: new Date(),
-    state: 'new',
-    status: 'available',
-    price: 5000000,
-    manufacturer: 'Toyota',
-    model: 'hilux',
-    bodyType: 'truck'
-  },
-  {
-    id: 5,
-    owner: 2, 
-    createdOn: new Date(),
-    state: 'new',
-    status: 'sold',
-    price: 5000000,
-    manufacturer: 'Toyota',
-    model: 'hilux',
-    bodyType: 'truck'
-  },
-  {
-    id: 6,
-    owner: 2, 
-    createdOn: new Date(),
-    state: 'used',
-    status: 'sold',
-    price: 4000000,
-    manufacturer: 'Toyota',
-    model: '530',
-    bodyType: 'SUV'
-  },
-];
-
-export default car;
