--- server/model/order.js
@@ -1,36 +1,0 @@
-const order = [
-  {
-    id: 1,
-    buyer: 2,
-    carId: 1,
-    price: 600000.23,
-    status: "pending",
-    priceOffered: 400000.58,
-  },
-  {
-    id: 2,
-    buyer: 2,
-    carId: 2,
-    price: 600000.23,
-    status: "pending",
-    priceOffered: 400000.58,
-  },
-  {
-    id: 3,
-    buyer: 2,
-    carId: 2,
-    price: 600000.23,
-    status: "approved",
-    priceOffered: 400000.58,  
-  },
-  {
-    id: 4,
-    buyer: 80,
-    carId: 2,
-    price: 600000.23,
-    status: "approved",
-    priceOffered: 400000.58,  
-  },
-];
-
-export default order;
