--- server/test/datas/car.js
@@ -1,52 +1,0 @@
-const car = [
-  {
-    id: 1,
-    owner: 1,
-    manufacturer: 'Toyota',
-    model: 'pathfinder',
-    price: 600000.23,
-    state: 'new',
-    status: 'sold',
-    newPrice: '45555252555',
-  },
-  {
-    id: 2,
-    owner: 1,
-    manufacturer: 'Toyota',
-    model: 'pathfinder',
-    newPrice: '45555252555',
-    price: '',
-    state: '',
-  },
-  {
-    id: 3,
-    owner: 2,
-    manufacturer: 'Toyota',
-    model: 'highlander',
-    price: 900000.23,
-    state: 'new',
-    status: 'sold',
-    newPrice: '45555252555',
-  },
-  {
-    id: 4,
-    owner: 1,
-    manufacturer: 'Toyota',
-    model: 'camry',
-    price: 5656256.23,
-    state: 'new',
-    status: 'sold',
-    newPrice: '45555252555',
-  },
-  {
-    id: 5,
-    manufacturer: 'Toyota',
-    model: 'camry',
-    price: 5656256.23,
-    state: 'new',
-    status: 'sold',
-    newPrice: '45555252555',
-  },
-];
-
-export default car;
