+++ server_data_structure/config/config.js
@@ -1,0 +1,8 @@
+import { Pool } from 'pg';
+import dotenv from 'dotenv';
+
+dotenv.config();
+const pool = new Pool({
+  connectionString: process.env.DB_CONNECTIONSTRING,
+});
+export default pool;
