+++ server_data_structure/controllers/autoController.js
@@ -1,0 +1,9 @@
+class autoController {
+  static welcome(req, res) {
+    return res.status(200).json({
+      message: 'Welcome to Auto-Mart',
+    });
+  }
+}
+
+export default autoController;
