+++ server_data_structure/helpers/passwordHash.js
@@ -1,0 +1,6 @@
+import bcrypt from 'bcryptjs';
+
+const hash = password => bcrypt.hashSync(password, 6);
+const unhash = (password, hashpassword) => bcrypt.compareSync(password, hashpassword);
+
+export { hash, unhash };
