+++ server_data_structure/helpers/token.js
@@ -7,4 +7,10 @@ const secretKey = process.env.secret;
 const token = payload => jwt.sign(payload, secretKey, {
   expiresIn: '1d',
 });
-export default { token };
+
+const verifyToken = tokenString => jwt.verify(tokenString, secretKey, (err, data) => {
+  if (err) return false;
+  return (data);
+});
+
+export { token, verifyToken };
