+++ server_data_structure/model/user.js
@@ -1,0 +1,32 @@
+const user = [
+  {
+    id: 1,
+    email: 'efejustin3@gmail.com',
+    firstName: 'Justin',
+    lastName: 'Igugu',
+    hashPassword: '$2a$06$IA8bQ5ZzEr4OJmXdL1Hz8O1ZLE7dinSSRFo0.poDt0.DsJUP7tmi6',
+    address: '15 omo avenue',
+    isAdmin: true
+  },
+  {
+    id: 2,
+    email: 'fejude3@gmail.com',
+    firstName: 'Nonso',
+    lastName: 'Amos',
+    hashPassword: '$2a$06$IA8bQ5ZzEr4OJmXdL1Hz8O1ZLE7dinSSRFo0.poDt0.DsJUP7tmi6',
+    address: '15 uloho avenue',
+    isAdmin: false
+  },
+  {
+    id: 3,
+    email: 'fejude3@rocketmail.com',
+    firstName: 'Kanayo',
+    lastName: 'Aass',
+    hashPassword: '$2a$06$IA8bQ5ZzEr4OJmXdL1Hz8O1ZLE7dinSSRFo0.poDt0.DsJUP7tmi6',
+    address: '15 uloho avenue',
+    isAdmin: false
+  },
+
+];
+
+export default user;
