+++ server_data_structure/test/datas/order.js
@@ -1,0 +1,25 @@
+const order = [
+  {
+    buyer: 1,
+    priceOffered: '400000.58',
+    newPriceOffered: '46525632612',
+  },
+  {
+    buyer: 2,
+    newPriceOffered: '46525632612',
+  },
+  {
+    buyer: 77,
+    newPriceOffered: '46525632612',
+  },
+  {
+    buyer: 77,
+    newPriceOffered: '46525632612',
+  },
+  {
+    buyer: 77,
+    newPriceOffered: '46525632612',
+  },
+];
+
+export default order;
